requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy
psycopg2-binary
pydantic[email]
nanoid
passlib
bcrypt
python-jose
python-multipart